+++ packages/rtk-rest-api/src/types/index.ts
@@ -26,6 +26,7 @@ export type {
   UpdateRoomPayload,
   CreateRoomPayload,
   RoomAssets,
+  RoomInvite,
 } from './room';
 export type { EventInvite, CreateEventInvitePayload } from './eventInvite';
 export {